simple_question_answering_app/requirements.txt
langchain
langchain-openai
langchain-community
openai
streamlit
python-dotenv